opencv/requirements.txt
opencv-contrib-python
pytesseract
absl-py